Add "smoke test" examples compilation CI workflow
On every push or pull request that affects library source or example files, and periodically, compile example
sketches for the specified boards.

Because this library doesn't currently have any example sketches, the Mouse built-in example sketches are used. This is
an artifact from a time when the Mouse functionality was part of the core library, rather than a standalone library.